tests/wpt/mozilla/tests/css/stylesheet_media_queries.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="myframe" src="iframe_for_media_queries.html" height="500px" width="500px" onLoad="func();">
</iframe>
<script>
var func = function() {
    var iframeTest = async_test("Media queries within stylesheets");
    var head = document.getElementById("myframe").contentWindow.document.getElementById("head");
    var before = getComputedStyle(head).getPropertyValue("background-color");
    document.getElementById("myframe").width = "300";
    setTimeout(function () {
        iframeTest.step(function () {
            var after = getComputedStyle(head).getPropertyValue("background-color");
            assert_true(before != after, "Recomputed values");
        });
        iframeTest.done();
    }, 1000);
}
</script>
